댓글 기능 구현

+++ articles/views.py
@@ -1,7 +1,7 @@
 from django.shortcuts import render, redirect
-from .forms import ArticleForm
-from .models import Article
-
+from .forms import ArticleForm,CommentForm
+from .models import Article, Comment
+from django.contrib.auth.decorators import login_required
 
 def index(request):
     articles = Article.objects.all()
@@ -13,25 +13,23 @@ def index(request):
 
 def detail(request, pk):
     article = Article.objects.get(pk=pk)
+    comments_form = CommentForm()
+    comments = article.comment_set.all()
     context = {
-        'article': article
+        'article': article,
+        'comments_form': comments_form,
+        'comments':comments,
     }
     return render(request, 'articles/detail.html', context)
 
-
-def new(request):
-    form = ArticleForm()
-    context = {
-        'form': form,
-    }
-    return render(request, 'articles/new.html', context)
-
-
+@login_required
 def create(request):
     if request.method == 'POST':
         form = ArticleForm(request.POST)
         if form.is_valid():
-            article = form.save()
+            article = form.save(commit=False)
+            article.user = request.user
+            article.save()
             return redirect('articles:detail', article.pk)
     else:
         form = ArticleForm()
@@ -41,35 +39,50 @@ def create(request):
     }
     return render(request, 'articles/new.html', context)
 
-
+@login_required
 def delete(request, pk):
     article = Article.objects.get(pk=pk)
-    article.delete()
+    if request.user == article.user:
+        article.delete()
     return redirect('articles:index')
 
-
-def edit(request, pk):
-    article = Article.objects.get(pk=pk)
-    form = ArticleForm(instance=article)
-    context = {
-        'form': form,
-        'article': article
-    }
-    return render(request, 'articles/edit.html', context)
-
-
+@login_required
 def update(request, pk):
     article = Article.objects.get(pk=pk)
-    if request.method == 'POST':
-        form = ArticleForm(request.POST, instance=article)
-        if form.is_valid():
-            form.save()
-            return redirect('articles:detail', article.pk)
+    if request.user == article.user:
+        if request.method == 'POST':
+            form = ArticleForm(request.POST, instance=article)
+            if form.is_valid():
+                form.save()
+                return redirect('articles:detail', article.pk)
+        else:
+            form = ArticleForm(instance=article)
     else:
-        form = ArticleForm(instance=article)
-
+        return redirect('articles:index')
     context = {
         'form': form,
         'article': article
     }
     return render(request, 'articles/edit.html', context)
+
+def comments_create(request, pk):
+    article = Article.objects.get(pk = pk)
+    comment_form = CommentForm(request.POST)
+    if comment_form.is_valid():
+        comment = comment_form.save(commit=False)
+        comment.article = article
+        comment.user = request.user
+        comment_form.save()
+        return redirect('articles:detail', article.pk)
+    context = {
+        'article':article,
+        'comment_form' : comment_form,
+    }
+    return render(request, 'articles/detail.html', context)
+
+@login_required
+def comments_delete(request, article_pk, comment_pk):
+    comment = Comment.objects.get(pk = comment_pk)
+    if request.user == comment.user:
+        comment.delete()
+    return redirect('articles:detail', article_pk)
